Added the SecurityService and a SHA256 implementation of it Refactored the ImageServiceImpl to use sha256 hashing for file naming for the images to avoid immediate duplicates and send back the existing copy Changed some of the basic exception throwbacks

+++ src/main/java/dev/ioannis/anemosparts/services/impl/ImageServiceImpl.java
@@ -1,6 +1,7 @@
 package dev.ioannis.anemosparts.services.impl;
 
 import dev.ioannis.anemosparts.services.ImageService;
+import dev.ioannis.anemosparts.services.SecurityService;
 import lombok.AllArgsConstructor;
 import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Value;
@@ -8,22 +9,28 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.naming.ServiceUnavailableException;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import java.security.NoSuchAlgorithmException;
 
 @Service
 public class ImageServiceImpl implements ImageService {
 
+    private SecurityService security;
+
     private String uploadDir;
     private String accessSrc;
 
-    public ImageServiceImpl(@Value("${app.upload.dir}") String uploadDir,  @Value("${app.access.src}") String accessSrc) {
+    public ImageServiceImpl(@Value("${app.upload.dir}") String uploadDir,  @Value("${app.access.src}") String accessSrc, SecurityService security) {
         this.uploadDir = uploadDir;
         this.accessSrc = accessSrc;
+
+        this.security = security;
 
         try{
             Files.createDirectories(Path.of(uploadDir));
@@ -35,7 +42,7 @@ public class ImageServiceImpl implements ImageService {
     }
 
     @Override
-    public String saveImage(MultipartFile file, String name) throws IOException {
+    public String saveImage(MultipartFile file) throws IOException, ServiceUnavailableException {
         if(file.isEmpty() || file.getSize() == 0 || file.getContentType().isEmpty()) {
             throw new IllegalArgumentException("Can't save an empty file");
         }
@@ -43,11 +50,21 @@ public class ImageServiceImpl implements ImageService {
             throw new IllegalArgumentException("Can't save an image of type " + file.getContentType());
         }
 
-        var fileName = name + '.' + FilenameUtils.getExtension(StringUtils.cleanPath(file.getOriginalFilename()));
 
-        Path path = Paths.get(uploadDir, fileName);
-        Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+        try {
+            String name = security.bytesToHash(file.getBytes());
 
-        return accessSrc + fileName;
+            var fileName = name + '.' + FilenameUtils.getExtension(StringUtils.cleanPath(file.getOriginalFilename()));
+
+            Path path = Paths.get(uploadDir, fileName);
+
+            if(Files.exists(path)) return accessSrc + fileName;
+
+            Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+
+            return accessSrc + fileName;
+        } catch (NoSuchAlgorithmException e) {
+            throw new ServiceUnavailableException("Can't save an image");
+        }
     }
 }
